fix(app): fixed style flash in search page
